+++ src/socialprofile/__init__.py
@@ -12,7 +12,7 @@ Release logic:
 
 import django
 
-__version__ = "1.23.34"
+__version__ = "1.23.35"
 __license__ = "MIT"
 __title__ = "django-errors"